requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.3

# Development
Flake8==3.9.2